Added wagon speed to config

+++ com/wurmonline/server/items/KingdomBanner.java
@@ -33,7 +33,7 @@ public class KingdomBanner {
         }
         builder.name(name + add + " banner", name + add + " banners", "An elegant symbol of allegiance and faith towards " + name);
         builder.descriptions("excellent", "good", "ok", "poor");
-        builder.itemTypes(new short[]{(short) 24, (short) 92, (short) 147, (short) 51, (short) 52, (short) 109, (short) 48, (short) 86, (short) 119, (short) 44, (short) 199, (short) 173});
+        builder.itemTypes(new short[]{24, 92, 147, 51, 52, 109, 48, 86, 119, 44, 199, 173});
         builder.imageNumber((short) 640);
         builder.combatDamage(0);
         builder.decayTime(9072000L);
